Separate REST frontend logic from business logic
This should make it more clear which parts belong to FLASK the the REST
service and which parts are just pure data broker logic.
